+++ Blocky/ModelObject.swift
@@ -13,5 +13,5 @@ protocol ModelObject {
 	var vertices: [Float] { get }
 	var vertexCount: Int { get }
 	
-	var indices: [Int] { get }
+	var indices: [UInt16] { get }
 }
